% Auto-generated: do not edit by hand
\name{dummyInclude}

\alias{dummyInclude}

\title{DummyInclude component}

\description{
MyTextInput is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
dummyInclude(id=NULL, label=NULL, value=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{label}{Character. The label used to show above of the input}

\item{value}{Character. The value used to set the default value for the input}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

